+++ CHANGELOG.md
@@ -60,6 +60,7 @@ The following changes have been completed, and will be included in the next tagg
     - made `URL.port` read-only
     - made `URL.protocol` read-only
     - made `URL.search` read-only
+    - made `URL.searchParams` read-only and side effect-free
   - added `isURLData()`
   - added `makeURL()`
   - added `mustBeURLData()`